Run the ELO refit as part of rcr enrich
